refactor(agents): reorganize agent actions and update import paths for better clarity
