+++ web-app/src/app/(app)/components/sidebar/components/agents-list.tsx
@@ -58,7 +58,6 @@ async function AgentsListContent() {
     userId,
     AgentListType.FAVORITE,
   );
-  console.log(agentList);
   const agentLists = [agentList];
 
   return (
